failing tests for liquidity simulation
Added tests for calling _accountLiquidity with  a non-zero borrowAmount
and a non-zero supplyAmount.

Had to modify the Exafin contract to allow for the entire token balance
to be withdrawn, which wasn't possible now. I think it'd be best to test
it explicitly but can't think of a way right now.

As far as I checked, having the trustedUnderlying token balance go down
to zero shouldn't be an issue because the `redeem` method is the only
place where it's used.