test: wipe out overdue debt
I only tested two iterations, but it got to the point where the
position's liquidity was greater than the owed amount, and ~3/4 of the
debt was repaid

This doesn't mean 3/4 of the collateral was seized, since the seize is
done at market value. I still have to look into that.