documented withdraw/deposits from 3rd parties
in particular, for the ETHFixedLender.
The purpose of this commit is to make visible the changes required now
that we realized having a custiodial ETHFixedLender would be highly
problematic

research notes:

erc2612 is the erc-20 permit standard
erc712 (confusing I know) is what's used to encode and sign arbitrary data, verifiably on the EVM
- [ ] Dapps are responsible for mitigating replay attacks (seeing the same message twice)
    - [ ] the chain id should be included to prevent replay attacks between different chains
- [ ] this gives a framework for signing _structs_ compatible with solidity
- [ ] using this implies adding a EIP712Domain struct to the signed data. All fields seem to be optional though
- [ ] the standard defines the web3 provider method for signing, and how it receives/returns data
- [ ] I'm not sure how would this be implemented in solidity. It looks like it'd be already implemented in a library or something. If I understood correctly it's not as simple as doing `keccak256(abi.encode(struct))`. That'd be dangerous since keccak256 fails injectiveness and/or determinism in only some cases.
Mon 10 Jan 2022 04:11:40 PM -03
update, it seems that what I want to achieve is possible with erc712. How easy it'll be to implement will depend on how much of the underlying thing I'll be able to abstract away
OZ's permit implementation seems simple enough, but I've yet to see the tests though

```
    function permit(
        address owner,
        address spender,
        uint256 value,
        uint256 deadline,
        uint8 v,
        bytes32 r,
        bytes32 s
    ) public virtual override {
        require(block.timestamp <= deadline, "ERC20Permit: expired deadline");

        bytes32 structHash = keccak256(abi.encode(_PERMIT_TYPEHASH, owner, spender, value, _useNonce(owner), deadline));

        bytes32 hash = _hashTypedDataV4(structHash);

        address signer = ECDSA.recover(hash, v, r, s);
        require(signer == owner, "ERC20Permit: invalid signature");

        _approve(owner, spender, value);
    }
```

... we'll need review from someone who actually knows their crypto
but as far as the ABI is concerned, I think we should only bother about adding the v,r,s parameters and then rebuild the signature struct on-chain
note: the EIP-712 from openzeppelin is still categorized as a draft