deleted now-unnecessary approve on the underlying.
the approve was necessary because previously we were doing
`SafeERC20.safeTransferFrom(address(this), recipient, amount)` and that
required an allowance to self to work. In a comment previous to this
one, and in a separate PR, we changed the `safeTransferFrom` to
`safeTransfer`, but didn't get rid of the approve. So we do it now.