assertions for rounding errors and their consequences
tried some low values for borrows trying to find some error in the
integrator that'd yield an inaccurate fee

managed to get a few scenarios:
- the utilization rate difference gets rounded down to zero, which'd
  cause the integrator to have a denominator of zero. An explicit check
  was added for that, yielding an INVALID_AMOUNT error
- the utilization rate difference was not zero, but the rate to be
  charged rounded down to zero
- the rate didnt get rounded down to zero, but the actual fee to be paid
  did
- the fee survived all roundings, and in the examples I tried, ended up
  being a valid value for the curve

takeaway is, rates can end up rounded down to zero if the supplied
liquidity is ~18 orders of magnitude greater than a single position. I
don't think that'll be a problem.

It'd be nice to have confirmation of this either theoretically from
francisco, or via fuzz testing