renamed Umax parameter
currently it was called targetUtilizationRate

this was wrong because the parameter is actually the max utilization
rate, or U_{max} in the docs, that is, the max utilization rate that's
valid, where the curve function shoots off into infinity.

the 'target utilization rate' is called U_{b} in the docs, and that's
the point separating the 'normal' from the 'leveraged' parts of the
curve. It is, however, not passed explicitly, and a consequence of A, B
and U_{max}.