renamed getFeeToBorrow -> getRateToBorrow
had to rename the previous getFeeToBorrow to simpsonIntegrator.

This was done since the value returned is not a fee (already scaled by
the amount) but just the rate, as a percentage that should be charged.

This is necessary since in borrows the amount is multiplied by the fee,
but in withdraws a division is made